telco-generator/requirements.txt
kafka-python
Faker
vastdb
pyarrow
ibis-framework

